EM/Battid framework: Fix temp_mon_range data type
BZ: 69934

temp_mon_ranges are defined as a 16 bit value in charging profile.
This needs to be reflected in the charging profile definition of
battid framework.

Reviewed-on: http://android.intel.com:8080/75617
Tested-by: Kallappa Manjanna, MadhukumarX <madhukumarx.kallappa.manjanna@intel.com>
Tested-by: cactus <cactus@intel.com>

+++ include/linux/power/battery_id.h
@@ -54,7 +54,7 @@ struct ps_pse_mod_prof {
 	u8 disch_tmp_ul;
 	u8 disch_tmp_ll;
 	/* number of temperature monitoring ranges */
-	u8 temp_mon_ranges;
+	u16 temp_mon_ranges;
 	struct ps_temp_chg_table temp_mon_range[BATT_TEMP_NR_RNG];
 	/* Lowest temperature supported */
 	short int temp_low_lim;
